Deployment Frequency Endpoint!
A variety of changes in this commit. The lambda folders have been renamed to use snake_case over kebab-case for python imports

There are new helper functions for the lambda, to make HTTP requests to Bitbucket and Jenkins.
The deployment frequency endpoint is currently hardcoded to be for all of time and for one singular CI/CD job